requirements.txt
# Core LangChain dependencies
langchain-core>=0.3.0
langchain-openai>=0.2.0

# LLM Providers
langchain-groq>=0.2.0
langchain-google-genai>=0.1.0

# Workflow orchestration
langgraph>=0.2.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# YAML configuration
pyyaml>=6.0.0

# Utilities
python-dateutil>=2.8.0

# Optional: For future providers (not yet implemented)
# langchain-anthropic>=0.2.0   # Claude support